+++ yachay_innopolis/src/innopolis/entities/Soliciestado.java
@@ -15,10 +15,10 @@ public class Soliciestado implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	@SequenceGenerator(name="SOLICIESTADO_IDSOLEST_GENERATOR", sequenceName="SEQ_SOLICIESTADO" , allocationSize=1)
+	@SequenceGenerator(name="SOLICIESTADO_IDSOLEST_GENERATOR", sequenceName="SEQ_SOLICIESTADO", allocationSize=1)
 	@GeneratedValue(strategy=GenerationType.SEQUENCE, generator="SOLICIESTADO_IDSOLEST_GENERATOR")
 	@Column(name="id_solest")
-	private Long idSolest;
+	private Integer idSolest;
 
 	private String estado;
 
@@ -29,11 +29,11 @@ public class Soliciestado implements Serializable {
 	public Soliciestado() {
 	}
 
-	public Long getIdSolest() {
+	public Integer getIdSolest() {
 		return this.idSolest;
 	}
 
-	public void setIdSolest(Long idSolest) {
+	public void setIdSolest(Integer idSolest) {
 		this.idSolest = idSolest;
 	}
