+++ yachay_innopolis/src/innopolis/entities/Tipoevento.java
@@ -15,10 +15,10 @@ public class Tipoevento implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	@SequenceGenerator(name="TIPOEVENTO_IDTIPOEVENTO_GENERATOR", sequenceName="SEQ_TIPOEVENTO"  , allocationSize=1)
+	@SequenceGenerator(name="TIPOEVENTO_IDTIPOEVENTO_GENERATOR", sequenceName="SEQ_TIPOEVENTO", allocationSize=1)
 	@GeneratedValue(strategy=GenerationType.SEQUENCE, generator="TIPOEVENTO_IDTIPOEVENTO_GENERATOR")
 	@Column(name="id_tipo_evento")
-	private Long idTipoEvento;
+	private Integer idTipoEvento;
 
 	private String descripcion;
 
@@ -31,11 +31,11 @@ public class Tipoevento implements Serializable {
 	public Tipoevento() {
 	}
 
-	public Long getIdTipoEvento() {
+	public Integer getIdTipoEvento() {
 		return this.idTipoEvento;
 	}
 
-	public void setIdTipoEvento(Long idTipoEvento) {
+	public void setIdTipoEvento(Integer idTipoEvento) {
 		this.idTipoEvento = idTipoEvento;
 	}
